allow user to specify a run containing only manual experiments
we achieve this by making the condition where num_wells == 0 meaningful
in qrandom.preprocess and sample. if the number of wells for a given
experiment is 0, the sampling loop will continue to the next experiment.
This avoids other errors that arose when num_wells was set to 0 but the
user did not comment out the experiment in the Template

this also required a modification to get_well_names: i.e. only read in
the manual_wells sheet if it is non-empty, based on the
rxndict['manual_wells']