lab config variables now controlled by Template file
this is achieved by using a dictionary in devconfig
which is keyed by the supported labs.

We also refactored utils.py into a directory
which contains the new code in globals.py
This code allows us to set and get global variables
across the repo. The motivation for this change
was to avoid having to pass the vardict or rxndict
everywhere where the user-specified lab was needed